begin work on shrink-to-fit 2nd pass layout mode
It should be noted that a major optimization was made impossible (or
at least significantly more difficult) due to the limitations of using
array-backed tree data structures. When in the 1st pass of shrink-to-fit and
a block has a length value for its width, then we should be able to
avoid processing its children until the 2nd pass. But trying to do
this with the current setup is problematic at the least.